refactor(view): console, web 패키지 구분
