+++ src/main/java/chess/WebApplication.java
@@ -10,8 +10,8 @@ import chess.dto.MoveResultDto;
 import chess.service.ChessJDBCDao;
 import chess.service.ChessService;
 import chess.service.DBChessServiceImpl;
-import chess.view.JsonTransformer;
-import chess.view.WebViewMapper;
+import chess.view.web.JsonTransformer;
+import chess.view.web.WebViewMapper;
 import com.google.gson.Gson;
 import java.util.HashMap;
 import java.util.Map;
