+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -2,8 +2,8 @@ package chess.domain.board;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
-import chess.view.PositionDto;
-import chess.view.WebViewMapper;
+import chess.dto.PositionDto;
+import chess.view.web.WebViewMapper;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
