+++ src/main/java/chess/domain/piece/Pawn.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import chess.domain.board.Board;
 import chess.domain.board.Position;
 import chess.domain.board.Rank;
-import chess.view.OutputView;
+import chess.view.console.OutputView;
 import java.util.List;
 
 public final class Pawn extends Piece {
