+++ src/main/java/chess/domain/piece/Rook.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.board.Board;
 import chess.domain.board.Position;
-import chess.view.OutputView;
+import chess.view.console.OutputView;
 import java.util.List;
 
 public final class Rook extends Piece {
