+++ src/main/java/chess/view/console/InputView.java
@@ -1,4 +1,4 @@
-package chess.view;
+package chess.view.console;
 
 import java.util.Arrays;
 import java.util.List;
