+++ src/main/java/chess/view/web/WebViewMapper.java
@@ -1,4 +1,4 @@
-package chess.view;
+package chess.view.web;
 
 import chess.domain.board.File;
 import chess.domain.board.Position;
@@ -11,6 +11,7 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
+import chess.dto.PositionDto;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
